requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-taggit==3.0.0
importlib-metadata==6.0.0
Markdown==3.4.1
psycopg2==2.9.5
sqlparse==0.4.3
zipp==3.11.0
